backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macc-backend"
version = "0.0.0"
description = "Local MACC modelling backend (FastAPI + SQLite)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.20",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
  "openai>=1.40",
  "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "types-python-jose>=3.3.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
